Implement robustness refactor: fix state persistence, add lifecycle, modularize
Phase 1: Fix state persistence bug (Issue #4)
- Rewrite restoreState() as thin adapter delegating to setState(), ensuring
  all setter side effects fire (observer activation, tick calculation,
  temperature state management)
- Exclude 'init' origin from pattr_state output to prevent default state
  overwriting saved data on load
- Clean up verbose investigation logging in getvalueof()/setvalueof()

Phase 2: Add initialization lifecycle management
- Add 'initialized' flag preventing pattr_state output during init/restore
- Set flag in init(), restoreState(), and setvalueof() for all paths
- Add [INIT-SEQUENCE] logging for empirical startup order documentation

Phase 3: Modularize into focused CommonJS modules
- Extract 9 modules: constants, utils, sequencer, observer-registry, state,
  instruments, commands, shuffle, temperature
- Temperature uses mixin pattern applied to SequencerDevice.prototype
- Consolidate 3 duplicated pitch offset calculations into _getCurrentPitchOffset()
- Rename lastAppliedValue -> lastParameterValue for clarity
- Fix stale notifydeleted() reference to removed transformations path
- Main file reduced from ~3013 to ~1741 lines

Phase 4: Documentation
- ADR-003: Robust State Restoration (Phases 1-2)
- ADR-004: Modularization (Phase 3)
- Updated CLAUDE.md key files table
- Added implementation log to robustness-refactor-plan.md
